+++ src/mvvm/awtk/binding_context_awtk.h
@@ -22,11 +22,14 @@
 #ifndef TK_BINDING_CONTEXT_AWTK_H
 #define TK_BINDING_CONTEXT_AWTK_H
 
+#include "base/widget.h"
 #include "mvvm/base/binding_context.h"
 
 BEGIN_C_DECLS
 
 binding_context_t* binding_context_create(void);
+
+ret_t binding_context_bind_simple(view_model_t* vm, widget_t* widget);
 
 END_C_DECLS