+++ src/mvvm/base/binding_context.c
@@ -44,17 +44,35 @@ ret_t binding_context_bind(binding_context_t* ctx, view_model_t* vm, void* widge
 }
 
 ret_t binding_context_update_to_view(binding_context_t* ctx) {
+  ret_t ret = RET_OK;
   return_value_if_fail(ctx != NULL && ctx->vt != NULL && ctx->vt->update_to_view != NULL,
                        RET_BAD_PARAMS);
 
-  return ctx->vt->update_to_view(ctx);
+  if(ctx->updating_view) {
+    return RET_BUSY;
+  } 
+
+  ctx->updating_view = TRUE;
+  ret = ctx->vt->update_to_view(ctx);
+  ctx->updating_view = FALSE;
+
+  return ret;
 }
 
 ret_t binding_context_update_to_model(binding_context_t* ctx) {
+  ret_t ret = RET_OK;
   return_value_if_fail(ctx != NULL && ctx->vt != NULL && ctx->vt->update_to_model != NULL,
                        RET_BAD_PARAMS);
 
-  return ctx->vt->update_to_model(ctx);
+  if(ctx->updating_model) {
+    return RET_BUSY;
+  } 
+
+  ctx->updating_model = TRUE;
+  ret = ctx->vt->update_to_model(ctx);
+  ctx->updating_model = FALSE;
+
+  return ret;
 }
 
 ret_t binding_context_destroy(binding_context_t* ctx) {