+++ src/tkc/darray.h
@@ -198,6 +198,17 @@ int32_t darray_count(darray_t* darray, void* ctx);
 ret_t darray_clear(darray_t* darray);
 
 /**
+ * @method darray_foreach
+ * 遍历元素。
+ * @param {darray_t*} darray 数组对象。
+ * @param {tk_visit_t} visit 遍历函数。
+ * @param {void*} ctx 遍历函数的上下文。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t darray_foreach(darray_t* darray, tk_visit_t visit, void* ctx);
+
+/**
  * @method darray_deinit
  * 清除全部元素，并释放elms。
  * @param {darray_t*} darray 数组对象。