+++ src/tkc/int_str.c
@@ -38,7 +38,7 @@ const char* int_str_value(const int_str_t* items, int32_t name) {
 
 int32_t int_str_name(const int_str_t* items, const char* value, int32_t defval) {
   const int_str_t* iter = items;
-  return_value_if_fail(items != NULL && value != NULL, NULL);
+  return_value_if_fail(items != NULL && value != NULL, defval);
 
   while (iter->value != NULL) {
     if (tk_str_eq(iter->value, value)) {